Move isolateAggregation to later phase
We cannot perform isolateAggregation as early as we currently do since
we may have user-defined aggregation functions that hasn't been resolved
yet.
