+++ frontend/src/main/scala/org/opencypher/v9_0/frontend/phases/ASTRewriter.scala
@@ -52,8 +52,6 @@ class ASTRewriter(rewriterSequencer: (String) => RewriterStepSequencer,
       normalizeSargablePredicates,
       enableCondition(normalizedEqualsArguments),
       addUniquenessPredicates,
-      isolateAggregation,
-      enableCondition(aggregationsAreIsolated),
       replaceLiteralDynamicPropertyLookups,
       namePatternComprehensionPatternElements,
       enableCondition(noUnnamedPatternElementsInPatternComprehension),
