+++ frontend/src/main/scala/org/opencypher/v9_0/frontend/phases/CompilationPhases.scala
@@ -30,6 +30,7 @@ object CompilationPhases {
       AstRewriting(sequencer, literalExtraction)
 
   def lateAstRewriting: Transformer[BaseContext, BaseState, BaseState] =
+    isolateAggregation andThen
     SemanticAnalysis(warn = false) andThen
       Namespacer andThen
       transitiveClosure andThen
